+++ src/plugins/azoth/plugins/sarin/filetransfermanager.h
@@ -66,7 +66,7 @@ namespace Sarin
 		void fileOffered (QObject*) override;
 
 		void gotFileControl (qint32, qint32, int);
-		void gotData (qint32, qint8, const QByteArray&);
+		void gotData (qint32, quint32, const QByteArray&);
 
 		// Internal signal to route file info requests from Tox thread to main thread.
 		void requested (int32_t, const QByteArray&, uint32_t, uint64_t, const QString&);
